optimize startup time
When launcher is started first time after booting, it takes about 2.5 s and
the screen is filled in two steps like version 1.9, but 500 ms faster.
Starting it a second time or later after booting, it takes only 1.5 s and
apears in one step most of the time unlike version 1.9 and again is 500 ms
faster.