Maximize window only for "Run" page to allow onboard keyboard to appear
On other pages show the window FullScreen, like before.

This is a workaround for a bug in onboard on-screen keyboard:
onboard stays behind a full screen window.
So, "Run" page of launcher would lack possibility to enter command on a touch
screen, which would make the page futile.
Reducing the launcher window to maximzed state lets the keyboard appear.
However, the taskbar gets visible, then.