fix : Suppress XMLHttpRequest warning and enable file read in future Qt versions
On bullseye (using Qt 5.15.2) the following warning is issued several
times on stderr, when raspad-launcher is started:
 XMLHttpRequest: Using GET on a local file is dangerous and will be disabled
 by default in a future Qt version.Set QML_XHR_ALLOW_FILE_READ to 1 if you
 wish to continue using this feature.

This is due to reading desktop files with the QML XMLHttpRequest object
using the GET method on file:// URLs in main.qml.
Setting the mentioned environment variable to 1 suppresses this warning,
causes no problem on buster (using Qt 5.11.3) and ensures that file://
access will still work with Qt 6.