Show window on startup only after all application files are loaded
The startup behaviour before was:
- show the main window with no icons in the GridView, first,
- load all application files (in about 450 ms), and
- finally show icons in GridView.

With this commit the startup behaviour is:
- load all application files, first,
- show the main window, completely with all icons.

This improves the "first impression" of raspad-launcher in the sense,
that the first glance that you catch at startup is already the complete
view. You don't have to watch the program getting ready.

The duration until the main window is completely shown and ready to use,
remains the same.